backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor>=3.3.0
pydantic>=2.0.0
python-multipart>=0.0.12
python-jose[cryptography]>=3.3.0
